Add a "development" run, which is not active in release builds.
This run is intended to be used for development of new functionality, in cases where manually triggering things is useful, without having to pause/unpause the supervisor/bot.

The supervisor needs to be refactored to properly support this "manual" type of run, as it checks inactivity of the player. As long as movement is done regularly (either manually, or by testing navigation/walking logic, etc.) the supervisor won't kill the run.